CTBase.jl

The CTBase.jl package is part of the control-toolbox ecosystem.

flowchart TD O(<a href='https://control-toolbox.org/OptimalControl.jl/stable/'>OptimalControl</a>) --> B(<a href='https://control-toolbox.org/OptimalControl.jl/stable/api-ctbase.html'>CTBase</a>) O --> D(<a href='https://control-toolbox.org/OptimalControl.jl/stable/api-ctdirect.html'>CTDirect</a>) O --> F(<a href='https://control-toolbox.org/OptimalControl.jl/stable/api-ctflows.html'>CTFlows</a>) F --> B D --> B style B fill:#FBF275

You may find in this package:

Types: the common types and associated methods if any.

Description: tools to manipulate tuples.

Differential geometry: tools to compute Lie brackets, Poisson brackets...

Exceptions.

Model: tools to modelise an optimal control problem in a functional manner.

Parser: tools to modelise an optimal control problem in an abstract manner.

Plot: the plot function extended from Plots.jl and utils.

Print: check how an optimal control problem is displayed.

Repl: control-toolbox repl.

Utils: auxiliary methods.

Here is the list of imported packages and exported functions.

CTBase.CTBase — Module

CTBase module.

Lists all the imported modules and packages:

Base

Core

DataStructures

DocStringExtensions

LinearAlgebra

MLStyle

Parameters

PrettyTables

Printf

ReplMaker

Unicode

List of all the exported names:

AbstractHamiltonian

AmbiguousDescription

Autonomous

BoundaryConstraint

CTCallback

CTCallbacks

CTException

Control

ControlConstraint

ControlLaw

Controls

Costate

Costates

DCostate

DState

Description

Dimension

Dynamics

FeedbackControl

Fixed

Hamiltonian

HamiltonianLift

HamiltonianVectorField

IncorrectArgument

IncorrectMethod

IncorrectOutput

Index

Lagrange

@Lie

Lie

Lift

Mayer

MixedConstraint

Model

Multiplier

NonAutonomous

NonFixed

NotImplemented

OptimalControlInit

OptimalControlModel

OptimalControlSolution

ParsingError

Poisson

PrintCallback

State

StateConstraint

States

StopCallback

Time

TimeDependence

Times

TimesDisc

UnauthorizedCall

Variable

VariableConstraint

VariableDependence

VectorField

__OCPModel

add

constraint

constraint!

constraint_type

constraints_labels

control!

ctNumber

ctVector

ct_repl

ct_repl_update_model

ctgradient

ctindices

ctinterpolate

ctjacobian

ctupperscripts

@def

dim_boundary_constraints

dim_control_constraints

dim_control_range

dim_mixed_constraints

dim_path_constraints

dim_state_constraints

dim_state_range

dim_variable_constraints

dim_variable_range

dynamics!

getFullDescription

get_priority_print_callbacks

get_priority_stop_callbacks

has_free_final_time

has_free_initial_time

has_lagrange_cost

has_mayer_cost

is_autonomous

is_fixed

is_max

is_min

is_time_dependent

is_time_independent

is_variable_dependent

is_variable_independent

nlp_constraints!

objective!

remove_constraint!

replace_call

set_AD_backend

state!

time!

variable!

∂ₜ

⋅

For the developers, here are the private methods.