Exceptions

Index

CTBase.AmbiguousDescription

CTBase.CTException

CTBase.ExtensionError

CTBase.IncorrectArgument

CTBase.IncorrectMethod

CTBase.IncorrectOutput

CTBase.NotImplemented

CTBase.ParsingError

CTBase.UnauthorizedCall

Documentation

CTBase.AmbiguousDescription — Type

struct AmbiguousDescription <: CTException

Exception thrown when the description is ambiguous / incorrect.

Fields

var::Tuple{Vararg{Symbol}}

CTBase.CTException — Type

abstract type CTException <: Exception

Abstract type for exceptions.

CTBase.ExtensionError — Type

mutable struct ExtensionError <: CTException

Exception thrown when an extension is not loaded but the user tries to call a function of it.

Fields

weakdeps::Tuple{Vararg{Symbol}}

CTBase.IncorrectArgument — Type

struct IncorrectArgument <: CTException

Exception thrown when an argument is inconsistent.

Fields

var::String

CTBase.IncorrectMethod — Type

struct IncorrectMethod <: CTException

Exception thrown when a method is incorrect.

Fields

var::Symbol

CTBase.IncorrectOutput — Type

struct IncorrectOutput <: CTException

Exception thrown when the output is incorrect.

Fields

var::String

CTBase.NotImplemented — Type

struct NotImplemented <: CTException

Exception thrown when a method is not implemented.

Fields

var::String

CTBase.ParsingError — Type

struct ParsingError <: CTException

Exception thrown for syntax error during abstract parsing.

Fields

var::String

CTBase.UnauthorizedCall — Type

struct UnauthorizedCall <: CTException

Exception thrown when a call to a function is not authorized.

Fields

var::String