CTFlows API

This is just a dump of CTFlows API documentation. For more details about CTFlows.jl package, see the documentation.

Index

CTFlows.Flow

CTFlows.Flow

Documentation

CTFlows.Flow — Method

Flow(
    ocp::OptimalControlModel{T, V},
    u_::Union{ControlLaw{T, V}, FeedbackControl{T, V}, Function},
    g_::Union{MixedConstraint{T, V}, StateConstraint{T, V}, Function},
    μ_::Union{Multiplier{T, V}, Function};
    alg,
    abstol,
    reltol,
    saveat,
    kwargs_Flow...
) -> CTFlows.OptimalControlFlow

Flow from an optimal control problem, a control function in feedback form, a state constraint and its associated multiplier in feedback form.

Example

julia> ocp = Model(autonomous=false)
julia> f = Flow(ocp, (t, x, p) -> p[1], (t, x, u) -> x[1] - 1, (t, x, p) -> x[1]+p[1])

Warning

The time dependence of the control function must be consistent with the time dependence of the optimal control problem. The dimension of the output of the control function must be consistent with the dimension usage of the control of the optimal control problem.

CTFlows.Flow — Method

Flow(
    ocp::OptimalControlModel{T, V},
    u_::Union{ControlLaw{T, V}, Function};
    alg,
    abstol,
    reltol,
    saveat,
    kwargs_Flow...
) -> CTFlows.OptimalControlFlow

Flow from an optimal control problem and a control function in feedback form.

Example

julia> f = Flow(ocp, (x, p) -> p)

Warning

The time dependence of the control function must be consistent with the time dependence of the optimal control problem. The dimension of the output of the control function must be consistent with the dimension usage of the control of the optimal control problem.