OptimalControl.jl (Private)

OptimalControl.jl is the root package of the control-toolbox ecosystem.

flowchart TD B(<a href='api-ctbase.html'>CTBase</a>) M(<a href='api-ctmodels.html'>CTModels</a>) P(<a href='api-ctparser.html'>CTParser</a>) O(<a href='api-optimalcontrol-dev.html'>OptimalControl</a>) D(<a href='api-ctdirect.html'>CTDirect</a>) F(<a href='api-ctflows.html'>CTFlows</a>) O --> D O --> M O --> F O --> P F --> M O --> B F --> B D --> B D --> M P --> M P --> B M --> B style O fill:#FBF275

Index

OptimalControl.__display

OptimalControl.clean

OptimalControl.version

Documentation

ExaModels.constraint! — Method

constraint!(ocp::CTModels.PreModel, type::Symbol; kwargs...)

See CTModels.constraint!.

OptimalControl.__display — Method

__display() -> Bool

Used to set the default display toggle. The default value is true.

OptimalControl.clean — Method

clean(d::Tuple{Vararg{Symbol}}) -> Tuple{Vararg{Symbol}}

When calling the function solve, the user can provide a description of the method to use to solve the optimal control problem. The description can be a partial description or a full description. The function solve will find the best match from the available methods, thanks to the function getFullDescription. Then, the description is cleaned by the function clean to remove the Symbols that are specific to OptimalControl.jl and so must not be passed to the solver. For instance, the Symbol :direct is specific to OptimalControl.jl and must be removed. It must not be passed to the CTDirect.jl solver.

OptimalControl.version — Method

Return the version of the current module as a string.

This function returns the version number defined in the Project.toml of the package to which the current module belongs. It uses @__MODULE__ to infer the calling context.

Example

julia> version()   # e.g., "1.2.3"