CTBase.jl

The CTBase.jl package is part of the control-toolbox ecosystem.

flowchart TD O(<a href='https://control-toolbox.org/OptimalControl.jl/stable/'>OptimalControl</a>) --> B(<a href='https://control-toolbox.org/OptimalControl.jl/stable/api-ctbase.html'>CTBase</a>) O --> D(<a href='https://control-toolbox.org/OptimalControl.jl/stable/api-ctdirect.html'>CTDirect</a>) O --> F(<a href='https://control-toolbox.org/OptimalControl.jl/stable/api-ctflows.html'>CTFlows</a>) F --> B D --> B style B fill:#FBF275

You may find in this package:

Description: tools to manipulate tuples.

Differential geometry: tools to compute Lie brackets, Poisson brackets...

Exceptions.

Initial guess: tools to define an initial guess.

Model: tools to modelise an optimal control problem in a functional manner.

Parser: tools to modelise an optimal control problem in an abstract manner.

Print: check how an optimal control problem is displayed.

Repl: control-toolbox repl.

Solution: optimal control solution.

Types: the common types and associated methods if any.

Utils: auxiliary methods.

Here is the list of imported packages and exported functions.

CTBase.CTBase — Module

CTBase module.

Lists all the imported modules and packages:

Base

Core

DataStructures

DocStringExtensions

LinearAlgebra

MLStyle

Parameters

PrettyTables

Printf

ReplMaker

Unicode

List of all the exported names:

AbstractHamiltonian

AmbiguousDescription

Autonomous

BoundaryConstraint

CTException

Control

ControlConstraint

ControlLaw

Controls

Costate

Costates

DCostate

DState

Description

Dimension

Dynamics

ExtensionError

FeedbackControl

Fixed

Hamiltonian

HamiltonianLift

HamiltonianVectorField

IncorrectArgument

IncorrectMethod

IncorrectOutput

Index

Lagrange

@Lie

Lie

Lift

Mayer

MixedConstraint

Model

Multiplier

NonAutonomous

NonFixed

NotImplemented

OptimalControlInit

OptimalControlModel

OptimalControlSolution

ParsingError

Poisson

State

StateConstraint

States

Time

TimeDependence

Times

TimesDisc

UnauthorizedCall

Variable

VariableConstraint

VariableDependence

VectorField

__OCPModel

add

boundary_constraints

boundary_constraints!

constraint

constraint!

constraint_type

constraints_labels

control

control!

control_components_names

control_constraints

control_constraints!

control_dimension

control_discretized

control_name

costate

costate!

costate_discretized

criterion

ctNumber

ctVector

ct_repl

ct_repl_update_model

ctgradient

ctindices

ctinterpolate

ctjacobian

ctupperscripts

@def

dim_boundary_constraints

dim_control_constraints

dim_control_range

dim_mixed_constraints

dim_path_constraints

dim_state_constraints

dim_state_range

dim_variable_constraints

dim_variable_range

dynamics

dynamics!

final_time

final_time_name

getFullDescription

has_free_final_time

has_free_initial_time

has_lagrange_cost

has_mayer_cost

infos

infos!

initial_time

initial_time_name

is_autonomous

is_fixed

is_max

is_min

is_time_dependent

is_time_independent

is_variable_dependent

is_variable_independent

iterations

iterations!

lagrange

mayer

message

message!

mixed_constraints

mixed_constraints!

model_expression

mult_boundary_constraints

mult_boundary_constraints!

mult_control_box_lower

mult_control_box_lower!

mult_control_box_upper

mult_control_box_upper!

mult_control_constraints

mult_control_constraints!

mult_mixed_constraints

mult_mixed_constraints!

mult_state_box_lower

mult_state_box_lower!

mult_state_box_upper

mult_state_box_upper!

mult_state_constraints

mult_state_constraints!

mult_variable_box_lower

mult_variable_box_lower!

mult_variable_box_upper

mult_variable_box_upper!

mult_variable_constraints

mult_variable_constraints!

nlp_constraints!

objective

objective!

remove

remove_constraint!

replace_call

set_AD_backend

state

state!

state_components_names

state_constraints

state_constraints!

state_dimension

state_discretized

state_name

stopping

stopping!

success

success!

time!

time_grid

time_grid!

time_name

variable

variable!

variable_components_names

variable_constraints

variable_constraints!

variable_dimension

variable_name

∂ₜ

⋅

For the developers, here are the private methods.