CTDirect API

This is just a dump of CTDirect API documentation. For more details about CTDirect.jl package, see the documentation.

Index

CTDirect.OCPSolutionFromDOCP

CTDirect.available_methods

CTDirect.directTranscription

CTDirect.getNLP

CTDirect.is_solvable

CTDirect.load_OCP_solution

CTDirect.save_OCP_solution

CTDirect.setDOCPInit

CTDirect.solve

CTDirect.solve

Documentation

CTDirect.OCPSolutionFromDOCP — Method

OCPSolutionFromDOCP(
    docp,
    docp_solution_ipopt
) -> OptimalControlSolution

Build OCP functional solution from DOCP vector solution (given as a GenericExecutionStats)

CTDirect.available_methods — Method

available_methods() -> Tuple{Tuple{Vararg{Symbol}}}

Return the list of available methods to solve the optimal control problem.

CTDirect.directTranscription — Method

directTranscription(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid
) -> CTDirect.DOCP

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

CTDirect.getNLP — Method

getNLP(docp::CTDirect.DOCP) -> Any

Extract the NLP problem from the DOCP

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Check if an OCP is solvable by the method solve.

CTDirect.load_OCP_solution — Function

load_OCP_solution(; ...) -> Any
load_OCP_solution(filename_prefix; format) -> Any

Load OCP solution in JLD2/JSON format

CTDirect.save_OCP_solution — Method

save_OCP_solution(
    sol::OptimalControlSolution;
    filename_prefix,
    format
)

Save OCP solution in JLD2/JSON format

CTDirect.setDOCPInit — Method

setDOCPInit(docp::CTDirect.DOCP, init) -> Any

Extract the NLP problem from the DOCP

CTDirect.solve — Method

solve(
    docp::CTDirect.DOCP;
    init,
    display,
    print_level,
    mu_strategy,
    kwargs...
) -> SolverCore.GenericExecutionStats

Solve a discretized optimal control problem DOCP

CTDirect.solve — Method

solve(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid,
    display,
    print_level,
    mu_strategy,
    kwargs...
) -> OptimalControlSolution

Solve an optimal control problem OCP by direct method