CTDirect.jl

The CTDirect.jl package is part of the control-toolbox ecosystem.

flowchart TD B(<a href='https://control-toolbox.org/OptimalControl.jl/stable/dev-ctbase.html'>CTBase</a>) M(<a href='https://control-toolbox.org/OptimalControl.jl/stable/dev-ctmodels.html'>CTModels</a>) P(<a href='https://control-toolbox.org/OptimalControl.jl/stable/dev-ctparser.html'>CTParser</a>) O(<a href='https://control-toolbox.org/OptimalControl.jl/stable/dev-optimalcontrol.html'>OptimalControl</a>) D(<a href='https://control-toolbox.org/OptimalControl.jl/stable/dev-ctdirect.html'>CTDirect</a>) F(<a href='https://control-toolbox.org/OptimalControl.jl/stable/dev-ctflows.html'>CTFlows</a>) O --> D O --> M O --> F O --> P F --> M O --> B F --> B D --> B D --> M P --> M P --> B M --> B style D fill:#FBF275

Index

CTDirect.DOCP

CTDirect.Gauss_Legendre_1

CTDirect.Gauss_Legendre_2

CTDirect.Gauss_Legendre_3

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.DOCP_constraints!

CTDirect.DOCP_initial_guess

CTDirect.DOCP_objective

CTDirect.IRK_dims

CTDirect.SolverInfos

CTDirect.__adnlp_backend

CTDirect.__disc_method

CTDirect.__display

CTDirect.__grid_size

CTDirect.__ipopt_linear_solver

CTDirect.__ipopt_mu_strategy

CTDirect.__ipopt_print_level

CTDirect.__knitro_print_level

CTDirect.__madnlp_linear_solver

CTDirect.__max_iterations

CTDirect.__ocp_init

CTDirect.__time_grid

CTDirect.__tolerance

CTDirect.add_nonzero_block!

CTDirect.available_methods

CTDirect.build_OCP_solution

CTDirect.build_OCP_solution

CTDirect.build_bounds

CTDirect.constraints_bounds!

CTDirect.direct_transcription

CTDirect.get_OCP_control_at_time_step

CTDirect.get_OCP_control_at_time_step

CTDirect.get_OCP_state_at_time_step

CTDirect.get_OCP_variable

CTDirect.get_lagrange_state_at_time_step

CTDirect.get_stagevars_at_time_step

CTDirect.get_time_grid

CTDirect.is_solvable

CTDirect.parse_DOCP_solution_dual

CTDirect.parse_DOCP_solution_primal

CTDirect.setPointConstraints!

CTDirect.setStepConstraints!

CTDirect.setStepConstraints!

CTDirect.setStepConstraints!

CTDirect.setStepConstraints!

CTDirect.setWorkArray

CTDirect.setWorkArray

CTDirect.setWorkArray

CTDirect.setWorkArray

CTDirect.setWorkArray

CTDirect.set_control_at_time_step!

CTDirect.set_initial_guess

CTDirect.set_optim_variable!

CTDirect.set_state_at_time_step!

CTDirect.solve

CTDirect.variables_bounds!

Documentation

Public

Private

CTDirect.DOCP — Type

Struct for discretized optimal control problem DOCP

Contains:

a copy of the original OCP

data required to link the OCP with the discretized DOCP

CTDirect.Gauss_Legendre_1 — Type

Implicit Midpoint discretization, formulated as a generic IRK (ie Gauss Legendre 1) For testing purpose only, use :midpoint instead (cf midpoint.jl) !

CTDirect.Gauss_Legendre_2 — Type

Gauss Legendre 2 discretization, formulated as a generic IRK

CTDirect.Gauss_Legendre_3 — Type

Gauss Legendre 3 discretization, formulated as a generic IRK

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{CTDirect.Euler}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{CTDirect.Midpoint}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{CTDirect.Trapeze}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{D<:CTDirect.Discretization}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{CTDirect.Euler}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{CTDirect.Midpoint}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{CTDirect.Trapeze}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{D<:CTDirect.Discretization}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_constraints! — Method

DOCP_constraints!(c, xu, docp::CTDirect.DOCP) -> Any

Compute the constraints C for the DOCP problem (modeled as LB <= C(X) <= UB).

CTDirect.DOCP_initial_guess — Function

DOCP_initial_guess(docp::CTDirect.DOCP) -> Vector{Float64}
DOCP_initial_guess(
    docp::CTDirect.DOCP,
    init::CTModels.Init
) -> Vector{Float64}

Build initial guess for discretized problem

CTDirect.DOCP_objective — Method

DOCP_objective(xu, docp::CTDirect.DOCP) -> Any

Compute the objective for the DOCP problem.

CTDirect.IRK_dims — Method

IRK_dims(
    dim_NLP_steps,
    dim_NLP_x,
    dim_NLP_u,
    dim_NLP_v,
    dim_path_cons,
    dim_boundary_cons,
    stage
) -> NTuple{5, Any}

Return the dimension of the NLP variables and constraints for a generic IRK discretizion, with the control taken constant per step (ie not distinct controls at time stages)

CTDirect.SolverInfos — Method

SolverInfos() -> Tuple{Int64, Float64, String, Symbol, Bool}

Retrieve convergence information (Ipopt version)

CTDirect.__adnlp_backend — Method

__adnlp_backend() -> Symbol

Used to set the default backend for AD in ADNLPModels. The default value is :optimized.

CTDirect.__disc_method — Method

__disc_method() -> Symbol

Used to set the default discretization method. The default value is trapeze.

CTDirect.__display — Method

__display() -> Bool

Used to set the default display toggle. The default value is true.

CTDirect.__grid_size — Method

__grid_size() -> Int64

Used to set the default grid size. The default value is 250.

CTDirect.__ipopt_linear_solver — Method

__ipopt_linear_solver() -> String

Used to set the default value of the linear solver of Ipopt for the direct method. The default value is mumps.

CTDirect.__ipopt_mu_strategy — Method

__ipopt_mu_strategy() -> String

Used to set the default value of the μ strategy of Ipopt for the direct method. The default value is adaptive.

CTDirect.__ipopt_print_level — Method

__ipopt_print_level() -> Int64

Used to set the default value of the print level of Ipopt for the direct method. The default value is 5.

CTDirect.__knitro_print_level — Method

__knitro_print_level() -> Int64

Used to set the default value of the print level of Knitro for the direct method. The default value is 3.

CTDirect.__madnlp_linear_solver — Method

__madnlp_linear_solver() -> String

Used to set the default value of the linear solver of MadNLP for the direct method. The default value is umfpack.

CTDirect.__max_iterations — Method

__max_iterations() -> Int64

Used to set the default maximum of iterations. The default value is 1000.

CTDirect.__ocp_init — Method

__ocp_init()

Used to set the default initial guess. The default value is nothing and will correspond to 0.1 for all variables.

CTDirect.__time_grid — Method

__time_grid()

Used to set the default time grid. The default value is nothing.

CTDirect.__tolerance — Method

__tolerance() -> Float64

Used to set the default tolerance. The default value is 1e-8.

CTDirect.add_nonzero_block! — Method

add_nonzero_block!(M, i_start, i_end, j_start, j_end; sym)

Add block of nonzeros elements to a sparsity pattern Format: boolean matrix (M) or index vectors (Is, Js) Includes a more compact method for single element case Option to add the symmetric block also (eg for Hessian) Note: independent from discretization scheme

CTDirect.available_methods — Method

available_methods(

) -> Tuple{Tuple{Symbol, Symbol}, Tuple{Symbol, Symbol}, Tuple{Symbol, Symbol}}

Return the list of available methods to solve the optimal control problem.

CTDirect.build_OCP_solution — Method

build_OCP_solution(
    docp,
    docp_solution
) -> CTModels.Solution{TimeGridModelType, TimesModelType, StateModelType, ControlModelType, VariableModelType, CostateModelType, Float64, DualModelType, CTModels.SolverInfos{Dict{Symbol, Any}}} where {TimeGridModelType<:CTModels.TimeGridModel, TimesModelType<:CTModels.TimesModel, StateModelType<:Union{CTModels.StateModelSolution{TS} where TS<:CTModels.var"#111#133", CTModels.StateModelSolution{TS} where TS<:CTModels.var"#112#134"}, ControlModelType<:Union{CTModels.ControlModelSolution{TS} where TS<:CTModels.var"#113#135", CTModels.ControlModelSolution{TS} where TS<:CTModels.var"#114#136"}, VariableModelType<:Union{CTModels.VariableModelSolution{Vector{Float64}}, CTModels.VariableModelSolution{Float64}}, CostateModelType<:Union{CTModels.var"#115#137", CTModels.var"#116#138"}, DualModelType<:(CTModels.DualModel{PC_Dual, BC_Dual, SC_LB_Dual, SC_UB_Dual, CC_LB_Dual, CC_UB_Dual, VC_LB_Dual, VC_UB_Dual} where {PC_Dual<:Union{Nothing, CTModels.var"#118#140", CTModels.var"#119#141"}, BC_Dual<:Union{Nothing, Vector{Float64}}, SC_LB_Dual<:Union{Nothing, CTModels.var"#121#143", CTModels.var"#122#144"}, SC_UB_Dual<:Union{Nothing, CTModels.var"#124#146", CTModels.var"#125#147"}, CC_LB_Dual<:Union{Nothing, CTModels.var"#127#149", CTModels.var"#128#150"}, CC_UB_Dual<:Union{Nothing, CTModels.var"#130#152", CTModels.var"#131#153"}, VC_LB_Dual<:Union{Nothing, Vector{Float64}}, VC_UB_Dual<:Union{Nothing, Vector{Float64}}})}

Build OCP functional solution from DOCP discrete solution (given as a SolverCore.GenericExecutionStats)

CTDirect.build_OCP_solution — Method

build_OCP_solution(docp; primal, dual, mult_LB, mult_UB)

Build OCP functional solution from DOCP discrete solution (given as a SolverCore.GenericExecutionStats)

CTDirect.build_bounds — Method

build_bounds(
    dim_var,
    dim_box,
    box_triplet
) -> Tuple{Any, Any}

Build full, ordered sets of bounds for state, control or optimization variables

CTDirect.constraints_bounds! — Method

constraints_bounds!(
    docp::CTDirect.DOCP
) -> Tuple{Vector{Float64}, Vector{Float64}}

Build upper and lower bounds vectors for the DOCP nonlinear constraints.

CTDirect.direct_transcription — Method

direct_transcription(
    ocp::Model,
    description...;
    grid_size,
    disc_method,
    time_grid,
    init,
    adnlp_backend,
    solver_backend,
    show_time,
    matrix_free
) -> Tuple{CTDirect.DOCP{_A, Model{TimesModelType, StateModelType, ControlModelType, VariableModelType, DynamicsModelType, ObjectiveModelType, ConstraintsModelType}} where {_A<:CTDirect.Discretization, TimesModelType<:CTModels.AbstractTimesModel, StateModelType<:CTModels.AbstractStateModel, ControlModelType<:CTModels.AbstractControlModel, VariableModelType<:CTModels.AbstractVariableModel, DynamicsModelType<:Function, ObjectiveModelType<:CTModels.AbstractObjectiveModel, ConstraintsModelType<:CTModels.AbstractConstraintsModel}, ADNLPModels.ADNLPModel{Float64, Vector{Float64}, Vector{Int64}}}

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

Arguments

ocp: optimal control problem as defined in CTModels

[description]: can specifiy for instance the NLP model and / or solver (:ipopt, :madnlp or :knitro)

Keyword arguments (optional)

grid_size: number of time steps for the discretized problem ([250])

disc_method: discretization method ([:trapeze], :euler, :euler_implicit, :midpoint, gauss_legendre_2, gauss_legendre_3)

time_grid: explicit time grid (can be non uniform)

init: info for the starting guess (values as named tuple or existing solution)

adnlp_backend: backend for automatic differentiation in ADNLPModels ([:optimized], :manual, :default)

show_time: (:true, [:false]) show timing details from ADNLPModels

CTDirect.get_OCP_control_at_time_step — Method

get_OCP_control_at_time_step(
    xu,
    docp::CTDirect.DOCP,
    i
) -> Any

Retrieve control variables at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps(+1), with convention u(tf) = U_N Vector output

CTDirect.get_OCP_control_at_time_step — Method

get_OCP_control_at_time_step(
    xu,
    docp::CTDirect.DOCP{CTDirect.Euler},
    i
) -> Any

Retrieve control variables at given time step from the NLP variables. Convention: see above for explicit / implicit versions Vector output

CTDirect.get_OCP_state_at_time_step — Method

get_OCP_state_at_time_step(
    xu,
    docp::CTDirect.DOCP,
    i
) -> Any

Retrieve state variables at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps+1 Vector output

CTDirect.get_OCP_variable — Method

get_OCP_variable(xu, docp::CTDirect.DOCP) -> Any

Retrieve optimization variables from the NLP variables. Convention: stored at the end, hence not dependent on the discretization method Vector output

CTDirect.get_lagrange_state_at_time_step — Method

get_lagrange_state_at_time_step(
    xu,
    docp::CTDirect.DOCP,
    i
) -> Any

Retrieve state variable for lagrange cost at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps+1 (no check for actual lagrange cost presence !)

CTDirect.get_stagevars_at_time_step — Method

get_stagevars_at_time_step(
    xu,
    docp::CTDirect.DOCP,
    i,
    j
) -> Any

Retrieve stage variables at given time step/stage from the NLP variables. Convention: 1 <= i <= dimNLPsteps(+1), 1 <= j <= s Vector output Note that passing correct indices is up to the caller, no checks are made here.

CTDirect.get_time_grid — Method

get_time_grid(xu, docp::CTDirect.DOCP) -> Any

Return time grid for variable time problems (times are then dependent on NLP variables)

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Check if an OCP is solvable by the method solve.

CTDirect.parse_DOCP_solution_dual — Method

parse_DOCP_solution_dual(
    docp,
    multipliers
) -> Tuple{Any, Any, Any}

Recover OCP costate and constraints multipliers from DOCP multipliers

CTDirect.parse_DOCP_solution_primal — Method

parse_DOCP_solution_primal(
    docp,
    solution;
    mult_LB,
    mult_UB
) -> Tuple{Any, Any, Any, NTuple{6, Any}}

Recover OCP primal variables from DOCP solution

CTDirect.setPointConstraints! — Method

setPointConstraints!(docp::CTDirect.DOCP, c, xu, v) -> Any

Set the boundary and variable constraints

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps (+1)

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Euler},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Midpoint},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setWorkArray — Method

setWorkArray(docp::CTDirect.DOCP, xu, time_grid, v) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Euler},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Midpoint},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.set_control_at_time_step! — Method

set_control_at_time_step!(
    xu,
    u_init,
    docp::CTDirect.DOCP,
    i
) -> Any

Set initial guess for control variables at given time step Convention: 1 <= i <= dimNLPsteps(+1)

CTDirect.set_initial_guess — Method

set_initial_guess(docp::CTDirect.DOCP, nlp, init) -> Any

Set initial guess in the DOCP

CTDirect.set_optim_variable! — Method

set_optim_variable!(xu, v_init, docp) -> Any

Set optimization variables in the NLP variables (for initial guess)

CTDirect.set_state_at_time_step! — Method

set_state_at_time_step!(
    xu,
    x_init,
    docp::CTDirect.DOCP,
    i
) -> Any

Set initial guess for state variables at given time step Convention: 1 <= i <= dimNLPsteps+1

CTDirect.solve — Method

solve(
    ocp::Model,
    description::Symbol...;
    display,
    grid_size,
    disc_method,
    time_grid,
    init,
    adnlp_backend,
    kwargs...
)

Solve an OCP with a direct method

Arguments

ocp: optimal control problem as defined in CTBase

[description]: can specifiy for instance the NLP model and / or solver (:ipopt, :madnlp or :knitro)

Keyword arguments (optional)

display: ([true], false) will disable output if set to false

grid_size: number of time steps for the discretized problem ([250])

disc_method: discretization method ([:trapeze], :midpoint, gauss_legendre_2)

time_grid: explicit time grid (can be non uniform)

init: info for the starting guess (values or existing solution)

adnlp_backend: backend for automatic differentiation in ADNLPModels ([:optimized], :manual, :default)

All further keywords are passed to the inner call of solve_docp

CTDirect.variables_bounds! — Method

variables_bounds!(
    docp::CTDirect.DOCP
) -> Tuple{Vector{Float64}, Vector{Float64}}

Build upper and lower bounds vectors for the DOCP variable box constraints.